refactor(php-transformer): freeze media-text contracts
