+++ php-transformer/src/HtmlToBlocks/Patterns/CoverPattern.php
@@ -29,6 +29,8 @@ use Throwable;
  */
 final class CoverPattern
 {
+    use PatternGateHelpersTrait;
+
     private CoverStyleResolver $styleResolver;
     private BackgroundImageExtractor $backgroundImageExtractor;
 
@@ -243,31 +245,6 @@ final class CoverPattern
         }
 
         return $count;
-    }
-
-    /**
-     * @param array<int, array<string, mixed>> $blocks
-     */
-    private function containsTextBearingBlock(array $blocks): bool
-    {
-        $textBearingNames = array( 'core/heading', 'core/paragraph', 'core/list', 'core/buttons', 'core/quote' );
-
-        foreach ( $blocks as $block ) {
-            if ( ! is_array($block) ) {
-                continue;
-            }
-
-            if ( in_array($block['blockName'] ?? null, $textBearingNames, true) ) {
-                return true;
-            }
-
-            $innerBlocks = $block['innerBlocks'] ?? array();
-            if ( is_array($innerBlocks) && $this->containsTextBearingBlock($innerBlocks) ) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     /**
